Changes to enable compilation in Windows with MSVC and MinGW32: - Undefine ERROR in Logger (already defined in Windows) - Undefine DEBUG in Prototype.h (defined by tinyxml2.h when targeting Debug version) - Fixed mkdir function parameters (slightly different in Windows) - Defined M_PI (unavailable in Windows) - Coded round function (unavailable in Windows) - Coded replacement for vasprintf (unavailable in Windows) - Coded equivalent replacement for coloured logger output - Included stdint.h to properly define data types like uint32_t - Fixed some errors in Path.cpp - Misc fixes

+++ source/Math/Rect.h
@@ -104,8 +104,8 @@ public:
 	inline T distanceX(const Rect<T> &r) const {
 		if (maxX() >= r.minX() && minX() <= r.maxX())
 			return 0;
-		return std::min<T>(	fabs(r.minX() - maxX()),
-							fabs(r.maxX() - minX()));
+		return std::min<T>(	fabs((float)r.minX() - maxX()),
+							fabs((float)r.maxX() - minX()));
 	}
 	
 	//Boolean operations
